Fix message handler test
